auth-features implemented

+++ routes/web.php
@@ -1,38 +1,38 @@
 <?php
 
+use App\Http\Controllers\Auth\AuthenticatedSessionController;
+use App\Http\Controllers\Auth\RegisteredUserController;
 use App\Http\Controllers\ItemController;
-use App\Http\Controllers\ProfileController;
 use Illuminate\Support\Facades\Route;
 
-Route::get('/', function () {
-    return view('welcome');
-});
-
-Route::get('/dashboard', function () {
-    return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
-
-Route::middleware('auth')->group(function () {
-    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
-});
 
 Route::prefix('items')->group(function(){
 
     Route::get('/', [ItemController::class, 'index'])->name('items-index');
 
-    Route::get('/create', [ItemController::class, 'create'])->name('items-create');
+    Route::get('/create', [ItemController::class, 'create'])->name('items-create')->middleware('auth');
 
-    Route::post('/', [ItemController::class, 'store'])->name('items-store');
+    Route::post('/', [ItemController::class, 'store'])->name('items-store')->middleware('auth');
 
-    Route::get('/show/{id}', [ItemController::class, 'show'])->name('items-show');
+    Route::get('/show/{id}', [ItemController::class, 'show'])->name('items-show')->middleware('auth');
 
-    Route::get('{id}/edit', [ItemController::class, 'edit'])->name('items-edit');
+    Route::get('{id}/edit', [ItemController::class, 'edit'])->name('items-edit')->middleware('auth');
 
-    Route::put('/{id}', [ItemController::class, 'update'])->name('items-update');
+    Route::put('/{id}', [ItemController::class, 'update'])->name('items-update')->middleware('auth');
 
-    Route::delete('/{id}', [ItemController::class, 'destroy'])->name('items-destroy');
+    Route::delete('/{id}', [ItemController::class, 'destroy'])->name('items-destroy')->middleware('admin');
+
+});
+
+Route::prefix('users')->group(function(){
+
+
+    Route::post('/login', [AuthenticatedSessionController::class, 'store'])->name('users-store');
+
+    Route::get('/logout', [AuthenticatedSessionController::class, 'destroy'])->name('users-destroy');
+
+    Route::post('/', [RegisteredUserController::class, 'store'])->name('users-create');
+
 });
